backend/requirements.txt
requests==2.31.0
openai==0.28.0
python-dotenv==1.0.0
moviepy==1.0.3
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
validators==0.20.0
pytest==7.4.0
pytest-cov==4.1.0
SQLAlchemy==2.0.23